added controller and jsp to choose countries client visited

+++ dao/src/main/java/dao/ClientDAO.java
@@ -6,5 +6,5 @@ import entity.Client;
 
 public interface ClientDAO extends ElementDAO<Client, Long> {
     List<String> getAllCountriesClientVisited(Integer clientId);
-    List<Integer> countVisas(Long clientId);
+    List<Integer> countVisas(Integer clientId);
 }
